Move extract ochre results to Analysis
